fix(Tree): fix the problem that tree nodes cannot be dragged
