+++ src/Tree/utils/index.ts
@@ -8,15 +8,6 @@ export { indentTreeNode } from './indentTreeNode';
 export { isSearching } from './isSearching';
 export { isExpand } from './isExpand';
 
-// Utility functions for tree node drag and drop.
-export {
-  getDragNodeKeys,
-  calDropNodePosition,
-  createDragTreeDataFunction,
-  createDragPreview,
-  removeDragPreview
-} from './treeDragDrop';
-
 // Utility functions for managing focusable items in a tree.
 export {
   getFocusableItems,
